Strengthen dense source validation and split tests
