+++ chutoro-core/src/error.rs
@@ -24,6 +24,9 @@ pub enum DataSourceError {
     /// Data source contained no rows.
     #[error("data source contains no rows")]
     EmptyData,
+    /// Data source rows must have positive dimension.
+    #[error("data source vectors must have positive dimension")]
+    ZeroDimension,
 }
 
 /// Error type produced when constructing or running [`Chutoro`].
