+++ chutoro-providers/dense/src/tests/mod.rs
@@ -1,0 +1,7 @@
+pub(crate) use super::{DenseMatrixProvider, DenseMatrixProviderError, DenseSource};
+
+mod errors;
+mod ingest;
+mod provider;
+mod source;
+mod support;
